+++ spec/spec_helper.rb
@@ -5,7 +5,3 @@ require 'awesome_print'
 # Requires supporting files with custom matchers and macros, etc,
 # in ./support/ and its subdirectories.
 Dir[File.expand_path(File.join(File.dirname(__FILE__),'support','**','*.rb'))].each {|f| require f}
-
-RSpec.configure do |config|
-  config.include Macros
-end