+++ HISTORY.md
@@ -1,6 +1,7 @@
 == 0.2.1 (2013-01-01)
 
 * Add new event fields.
+* Handle multi-line values.
 
 == 0.2 (2013-01-01)